* 'hotfix/fixbug' of github.com:zhiyicx/thinksns-plus-android:
  (#1329) code fix this
  (#1182) code 注册协议
  (#1324) code 网页对齐方式

* 'hotfix/fixbug' of github.com:zhiyicx/thinksns-plus-android:
  (#1329) code fix this
  (#1182) code 注册协议
  (#1324) code 网页对齐方式

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/beans/DynamicDetailBeanV2.java
@@ -720,7 +720,7 @@ public class DynamicDetailBeanV2 extends BaseListBean implements Parcelable, Ser
         dest.writeValue(this.feed_mark);
         dest.writeByte(this.has_digg ? (byte) 1 : (byte) 0);
         dest.writeByte(this.has_collect ? (byte) 1 : (byte) 0);
-        dest.writeDouble(this.amount);
+        dest.writeLong(this.amount);
         dest.writeTypedList(this.likes);
         dest.writeByte(this.paid ? (byte) 1 : (byte) 0);
         dest.writeTypedList(this.images);
